pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Sergey Radyukhin <22dollars@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.4"
python-dotenv = "^0.21.0"
django-bootstrap4 = "^22.3"
dj-database-url = "^1.2.0"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
ipython = "^8.7.0"
django-extensions = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
